Lannister/api_src/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
python-dotenv==0.18.0
pytest==7.1.2
datetime
loguru==0.6.0
datetime
djangorestframework==3.13.1